Hi all,
I promised Geert Stappers to work on a Dutch translation of the
conglomerate .po-file.

The translation is not usable yet. I send it in anyway because I'll be
really busy for the next few weeks.
I hope you'll be able to use at least some of this.

Sorry about the rough edges.

Sincerely,

Mark Van den Borre
Classical guitarist, Digital Freedom specialist