Remove use of deprecated Gnome libraries
See bug #656279.

This patch removes the dependency on libgnome, libgnomeui, and
libbonobo. Some specifics:
- Removes GnomeOption/POPT parsing in favor of g_option_...
- Removes use of GnomeProgram. Doesn't replace it with anything yet,
  since GApplication isn't supported well on all platforms.
- gnome_program_locate_file() now looks in DATADIR/conglomerate
- Removes use of GnomeDruid in favor of GtkAssistant
- Removes use of various functions and macros in favor of their modern
  counterparts.

The only thing not replaced is the use of GnomeEntry in
cong-edit-find-and-replace.c. GnomeEntry saves previous text values
across runs and adds them to a completion list. The proper way to do
this without libgnomeui is to save the values in one's own configuration
file, but that should be done when the application is ported to
GSettings.